off_t for file size > 4G

+++ machine.virt/system/drivers/keybd/keybd.c
@@ -78,7 +78,7 @@ int get_key_code(char *buf)
 }
 
 static int keybd_read(vdevice_t* dev, int fd, int from_pid, fsinfo_t *info,
-                      void *buf, int size, int offset, void *p)
+                      void *buf, int size, off_t offset, void *p)
 {
     (void)dev;
     (void)fd;
