+++ machine.virt/system/drivers/powerd/powerd.c
@@ -8,7 +8,7 @@
 #include <ewoksys/dma.h>
 
 static int power_read(vdevice_t* dev, int fd, int from_pid, fsinfo_t* info,
-        void* buf, int size, int offset, void* p) {
+        void* buf, int size, off_t offset, void* p) {
     (void)dev;
     (void)fd;
     (void)from_pid;
