+++ system/basic/drivers/logd/logd.c
@@ -14,7 +14,7 @@ static int log_read(vdevice_t* dev,
         fsinfo_t* info,
         void* buf,
         int size,
-        int offset,
+        off_t offset,
         void* p) {
 
     (void)dev;
@@ -64,7 +64,7 @@ static int log_write(vdevice_t* dev,
         fsinfo_t* info,
         const void* buf,
         int size,
-        int offset,
+        off_t offset,
         void* p) {
 
     (void)dev;
