+++ system/basic/drivers/nulld/nulld.c
@@ -12,7 +12,7 @@ static int null_read(vdevice_t* dev,
         fsinfo_t* info,
         void* buf,
         int size,
-        int offset,
+        off_t offset,
         void* p) {
 
     (void)dev;
@@ -32,7 +32,7 @@ static int null_write(vdevice_t* dev,
         fsinfo_t* info,
         const void* buf,
         int size,
-        int offset,
+        off_t offset,
         void* p) {
 
     (void)dev;
