+++ system/basic/libc/libewoksys/ewoksys/include/ewoksys/devcmd.h
@@ -13,8 +13,8 @@ extern int dev_stat(int dev_pid, fsinfo_t* info, node_stat_t* stat);
 extern int dev_create(int dev_pid, fsinfo_t* info_to, fsinfo_t* info);
 extern int dev_unlink(int dev_pid, ewokos_addr_t node, const char* fname);
 extern int dev_open(int dev_pid, int fd, fsinfo_t* info, int oflag);
-extern int dev_read(int dev_pid, int fd, fsinfo_t* info, int32_t offset, void* buf, uint32_t size);
-extern int dev_write(int dev_pid, int fd, fsinfo_t* info, int32_t offset, const void* buf, uint32_t size);
+extern int dev_read(int dev_pid, int fd, fsinfo_t* info, off_t offset, void* buf, uint32_t size);
+extern int dev_write(int dev_pid, int fd, fsinfo_t* info, off_t offset, const void* buf, uint32_t size);
 /* lifecycle hooks for the per-fd persistent transfer-buffer cache */
 extern void dev_io_on_close(int fd);
 extern void dev_io_on_fork(void);
