+++ system/basic/libc/libgloss/syscalls.c
@@ -309,14 +309,14 @@ off_t
 _lseek (int fd, off_t offset, int whence)
 {
     if(whence == SEEK_CUR) {
-        int cur = vfs_tell(fd);
+        off_t cur = vfs_tell(fd);
         if(cur < 0)
             cur = 0;
         offset += cur;
     }
     else if(whence == SEEK_END) {
         fsinfo_t info;
-        int cur = 0;
+        off_t cur = 0;
         if(vfs_get_by_fd(fd, &info) == 0)
             cur = info.stat.size;
         offset += cur;
