+++ system/basic/sys/init/readfs.c
@@ -21,7 +21,7 @@ static int32_t ext_sd_read_blocks(int32_t block, void* buf, uint32_t count) {
 
 /* ext3 path through the ext3 library: the journal is recovered on init and
  * the fs is committed/closed cleanly on quit; only the read API is used. */
-static void* read_ext3(const char* fname, int32_t* size) {
+static void* read_ext3(const char* fname, off_t* size) {
     ext3_t ext3;
     if(ext3_init_ex2(&ext3, sd_read, ext_sd_read_blocks, sd_write, NULL,
             sd_flush, SD_BUFFER_SIZE) != 0) {
@@ -34,7 +34,7 @@ static void* read_ext3(const char* fname, int32_t* size) {
 
 /* same contract as the kernel's read_fs: probe the partition and use the
  * matching reader (ext3 preferred when a usable journal is present) */
-void* read_fs(const char* fname, int32_t* size) {
+void* read_fs(const char* fname, off_t* size) {
     return read_ext3(fname, size);
     //return read_ext2(fname, size);
 }
