+++ system/xwin/apps/xterm/xterm.cc
@@ -514,7 +514,7 @@ static int console_write(vdevice_t* dev, int fd,
 		fsinfo_t* info,
 		const void* buf,
 		int size,
-		int offset,
+		off_t offset,
 		void* p) {
 	(void)dev;
 	(void)fd;
@@ -544,7 +544,7 @@ static int console_write(vdevice_t* dev, int fd,
 }
 
 static int console_read(vdevice_t* dev, int fd, int from_pid, fsinfo_t* info,
-		void* buf, int size, int offset, void* p) {
+		void* buf, int size, off_t offset, void* p) {
 	(void)dev;
 	(void)fd;
 	(void)offset;
