+++ system/xwin/drivers/vjoystickd/vjoystickd.c
@@ -133,7 +133,7 @@ static int vjoystick_read(vdevice_t* dev, int fd,
         fsinfo_t* info,
         void* buf,
         int size,
-        int offset,
+        off_t offset,
         void* p) {
 
     (void)dev;
